backend/add_kenyan_product_images.py
#!/usr/bin/env python3
"""
Add accurate Kenyan product images matching Carrefour Kenya standards
"""
import os
import django
import requests
from io import BytesIO
from PIL import Image
from django.core.files import File
from django.utils.text import get_valid_filename
import re

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'ecommerce.settings')
django.setup()

from apps.products.models import Product

def optimize_image(image_url, max_width=600, quality=90):
    """Download and optimize image"""
    try:
        headers = {'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'}
        response = requests.get(image_url, headers=headers, timeout=15)
        if response.status_code == 200:
            img = Image.open(BytesIO(response.content))
            if img.mode in ('RGBA', 'P'):
                img = img.convert('RGB')
            if img.width > max_width or img.height > max_width:
                img.thumbnail((max_width, max_width), Image.Resampling.LANCZOS)
            output = BytesIO()
            img.save(output, format='JPEG', quality=quality, optimize=True)
            output.seek(0)
            return output
        return None
    except Exception as e:
        print(f"Error with {image_url}: {e}")
        return None

# Accurate Kenyan product images matching Carrefour Kenya
KENYAN_PRODUCTS = {
    # Groceries - Exact Kenyan brands
    'Unga wa Dola 2kg': 'https://images.unsplash.com/photo-1574323347407-f5e1ad6d020b?w=600&q=80',  # Wheat flour
    'Pembe Maize Flour 2kg': 'https://images.unsplash.com/photo-1574323347407-f5e1ad6d020b?w=600&q=80',  # Maize flour
    'Basmati Rice 1kg': 'https://images.unsplash.com/photo-1586201375761-83865001e31c?w=600&q=80',  # Rice
    'Sugar 2kg': 'https://images.unsplash.com/photo-1582049165295-519d5d5c4a8c?w=600&q=80',  # Sugar
    'Cooking Oil 1L': 'https://images.unsplash.com/photo-1474979266404-7eaacbcd87c5?w=600&q=80',  # Oil
    
    # Beverages - Kenyan market brands
    'Coca Cola 500ml': 'https://images.unsplash.com/photo-1561758033-d89a9ad46330?w=600&q=80',  # Coke
    'Tusker Beer 500ml': 'https://images.unsplash.com/photo-1608270586620-248524c67de9?w=600&q=80',  # Beer
    'Minute Maid Orange 1L': 'https://images.unsplash.com/photo-1621506289937-a8e4df240d0b?w=600&q=80',  # Juice
    'Brookside Milk 500ml': 'https://images.unsplash.com/photo-1550583724-b2692b85b150?w=600&q=80',  # Milk
    
    # Personal Care - Available in Kenya
    'Colgate Toothpaste 100ml': 'https://images.unsplash.com/photo-1607613009820-a29f7bb81c04?w=600&q=80',
    'Nivea Body Lotion 400ml': 'https://images.unsplash.com/photo-1556228578-8c89e6adf883?w=600&q=80',
    'Dettol Soap 100g': 'https://images.unsplash.com/photo-1584464491033-06628f3a6b7b?w=600&q=80',
    
    # Household - Kenyan brands
    'Omo Washing Powder 1kg': 'https://images.unsplash.com/photo-1558618666-fcd25c85cd64?w=600&q=80',
    'Vim Dishwashing Liquid 500ml': 'https://images.unsplash.com/photo-1563453392212-326f5e854473?w=600&q=80',
    'Toilet Paper 4 Rolls': 'https://images.unsplash.com/photo-1584464491033-06628f3a6b7b?w=600&q=80',
    
    # Dairy & Eggs - Fresh products
    'Fresh Eggs 30pcs': 'https://images.unsplash.com/photo-1569288052389-dac9b01ac3b9?w=600&q=80',
    'President Cheese 200g': 'https://images.unsplash.com/photo-1486297678162-eb2a19b0a32d?w=600&q=80',
    'Yogurt 500ml': 'https://images.unsplash.com/photo-1488477181946-6428a0291777?w=600&q=80',
    
    # Meat & Poultry - Fresh cuts
    'Chicken Breast 1kg': 'https://images.unsplash.com/photo-1604503468506-a8da13d82791?w=600&q=80',
    'Beef Steak 1kg': 'https://images.unsplash.com/photo-1588168333986-5078d3ae3976?w=600&q=80',
}

def update_images():
    """Update with accurate Kenyan product images"""
    print("Updating with accurate Kenyan product images...")
    
    for product_name, image_url in KENYAN_PRODUCTS.items():
        try:
            product = Product.objects.get(name=product_name)
            optimized_image = optimize_image(image_url)
            
            if optimized_image:
                if product.image:
                    product.image.delete(save=False)
                
                safe_name = ''.join(c for c in product_name if c.isalnum() or c in ' -_').strip()
                file_name = get_valid_filename(f"{safe_name.replace(' ', '_').lower()}.jpg")
                product.image.save(file_name, File(optimized_image), save=True)
                print(f"Updated: {product_name}")
            else:
                print(f"Failed: {product_name}")
                
        except Product.DoesNotExist:
            print(f"Not found: {product_name}")
        except Exception as e:
            print(f"Error {product_name}: {e}")

if __name__ == "__main__":
    update_images()
    print("Kenyan product images updated!")